LayoutTests/imported/w3c/web-platform-tests/resize-observer/change-layout-in-error-expected.txt

CONSOLE MESSAGE: ResizeObserver loop completed with undelivered notifications.
CONSOLE MESSAGE: ResizeObserver loop completed with undelivered notifications.

PASS Changing layout in window error handler should not result in lifecyle loop when resize observer loop limit is reached. 
